refactor: move editor into tools
