--- core/editor/__init__.py
@@ -1,3 +1,0 @@
-from .patch import CodePatcher
-from .read import CodeReader
-from .write import CodeWriter
