+++ core/tools/editor/patch.py
