+++ core/tools/editor/read.py
