+++ core/tools/editor/write.py
